hotfix: validaciones por rol y colspan modificado

+++ server/controllers/StockController.js
@@ -41,6 +41,10 @@ export default class StockController {
    */
   static async create(req, res) {
     try {
+      if (!user || !isEqual(user.role, "admin")) {
+        return res.status(403).json();
+      }
+
       const { name, quantity, price, category, supplyDate } = req.body;
       const stock = await Stock.create({
         name,
@@ -61,11 +65,16 @@ export default class StockController {
    */
   static async update(req, res) {
     try {
+      if (!user || isEqual(user.role, "client")) {
+        return res.status(403).json();
+      }
       const id = req.params.id;
-      if (!id) return res.status(400).send();
-      const body = req.body;
-      if (!body) return res.status(404).send();
-      await Stock.update(body, { where: { id } });
+      if (!id || !req.body) return res.status(400).send();
+      const { name, quantity, price, category, supplyDate } = req.body;
+      await Stock.update(
+        { name, quantity, price, category, supplyDate },
+        { where: { id } }
+      );
       res.status(200).send();
     } catch (e) {
       res.status(500).send();
@@ -78,6 +87,9 @@ export default class StockController {
    */
   static async delete(req, res) {
     try {
+      if (!user || !isEqual(user.role, "admin")) {
+        return res.status(403).json();
+      }
       const id = req.params.id;
       if (!id) return res.status(400).send();
       await Stock.destroy({ where: { id } });
